requirements.txt
flask>=1.1.1
flask-login>=0.5.0
Flask-SQLAlchemy>=2.4.1
Flask-WTF>=0.14.3
itsdangerous>=1.1.0
Jinja2>=2.11.1
MarkupSafe>=1.1.1
python-dotenv>=0.12.0
SQLAlchemy>=1.3.15
Werkzeug>=1.0.1
WTForms>=2.2.1
flake8
pytest
email_validator
sqlalchemy_utils
requests
python-dateutil
Flask-Migrate
onetimepass
pyqrcode
timecop
cerberus